Add FSMs for CENs and CINs; introduce executive
The lev_executive module is the entry point to the system. It is used
by the leviathan_rest as a callback module for handling requests.

The FSMs are spawned per a CEN/CIN and a host, that particular CEN/CIN
spans over. They react on the status changes in Dobby identifiers tied
to the CENs and CINs.

The CEN/CIN LMs are distributed over the cluster by Distributed Erlang.
Each Erlang nodes manages its authoritative store, which currently has
the same content on each node.